refactor: class-tag read-point hygiene from the rotation 497 audit
The audit of every runtime `struct_layout_lookup` / OBJ_CLASS_TAG_OFF
read point (30 files) found no unguarded site; these are the audit's
zero-risk residuals. define_struct.rs and json_stringify/composites.rs
spelled the class-tag offset as a bare `8`, invisible to a read-point
sweep — named constant. member_get.rs declared `__torajs_struct_layout_
lookup` / `_field_find` externs no call site used and a doc that
claimed a walk lived there — removed. instanceof_any.rs read the u32
class-tag slot as i64 (the only 8-byte reader; the pad would fold into
the comparison if it were ever non-zero) — u32 read, widened to compare.

+++ crates/torajs-rc/src/instanceof_any.rs
@@ -170,8 +170,10 @@ pub unsafe extern "C" fn __torajs_instanceof_class_any_tag(v: i64, expected_tag:
         }
         return false;
     }
-    let class_tag = unsafe { *((ptr as *const u8).add(OBJ_CLASS_TAG_OFF) as *const i64) };
-    class_tag == expected_tag
+    // The slot is a u32 (every other read point agrees); reading 8
+    // bytes would fold the pad after it into the comparison.
+    let class_tag = unsafe { *((ptr as *const u8).add(OBJ_CLASS_TAG_OFF) as *const u32) };
+    i64::from(class_tag) == expected_tag
 }
 
 #[cfg(not(test))]
